refactor: externalize taunts and update code
Truthfully, I don't really remember the last changes I made to this
code, since it was a while ago, but I need to wipe this machine, so I'm
commiting and pushing the code.

I have saved the TOKEN and ADMIN_ID environment variables for Cato in my
Dashlane account for future recovery.  They should be placed in the .env
file to begin using this project again.